Better support for mysql Enums
Instead of mapping all enums to be treated as "strings", I created a new pseudo type so that DBAL Dotrine is happy. Since we're only using super basic functionality of the DBAL library, it actually works fine (well, better) this way.

This will automatically determine what values can go in a mysql "enum" field, and create the factory appropriately.